% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{bandlimited_density}
\alias{bandlimited_density}
\title{One-Step Estimator of a Bandlimited Counterfactual Density}
\usage{
bandlimited_density(
  W,
  A,
  Y,
  ngrid = 1000,
  num_fold = 2,
  alpha = 0.05,
  num_boot = 2000,
  b = 10
)
}
\arguments{
\item{W}{data frame with observations in the rows and baseline covariates in the columns.}

\item{A}{binary treatment vector, where 1 indicates that an individual was treated.}

\item{Y}{real-valued outcome.}

\item{ngrid}{number of grid points used in approximation procedure. Selecting larger values of ngrid should yield better performance, but slower runtime.}

\item{num_fold}{Number of folds used during cross-fitting. Should be at least 2.}

\item{alpha}{1-alpha is the desired level of confidence set. Can be a single number between 0 and 1, or a vector of such numbers.}

\item{num_boot}{Number of bootstrap replications used to estimate appropriate L2 radius of confidence set.}

\item{b}{Upper bound on the support of the Fourier transform of the counterfactual density function. If this Fourier transform has support outside of \code{[-b,b]}, then the estimand is instead the b-bandlimiting of this function.}
}
\value{
a list containing
\item{onestep_fun}{A function taking as input a vector of outcome values and outputting one-step estimates of the b-bandlimited counterfactual density function at those values.}
\item{plugin_fun}{Same as the above, but the estimates are obtained based on a plug-in estimator. Provided for comparison purposes only --- \code{onestep_fun} is the preferred estimator.}
\item{L2_radius_sq}{A vector of the squared L2 radiuses of alpha-level confidence sets for the b-bandlimited counterfactual density function. These confidence sets consist of all b-bandlimited functions contained in L2 balls centered at \code{onestep_fun} with the specified radii.}
}
\description{
This function estimates a bandlimited counterfactual density function under treatment A=1.
}
\details{
If the counterfactual density function is not bandlimited at the specified value \code{b}, then a \code{b}-bandlimiting of the density function is estimated instead. See the reference for details.

All needed nuisances are estimated using the ranger package.
}
\examples{
# sample size
n = 500

# simulate data
dat = sim_data_bandlimited(n)
W = dat$W
A = dat$A
Y = dat$Y

bl_out = bandlimited_density(W,A,Y,ngrid=500,num_fold=2,num_boot=2000,alpha=seq(0.01,0.99,by=0.01),b=2)
ygrid = seq(-15,15,by=0.01)
plot(ygrid,bl_out$onestep_fun(ygrid),xlab='y',ylab='Counterfactual Density Estimate',type='l')
}
\references{
A. Luedtke and I. Chung, ``One-Step Estimation of Differentiable Hilbert-Valued Parameters,'' \emph{arXiv}, 2023.
}
