% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{cv_density}
\alias{cv_density}
\title{Regularized one-step estimator of counterfactual density function}
\usage{
cv_density(
  W,
  A,
  Y,
  ngrid = 500,
  dens_init_method = "ranger",
  maxK_legendre = 16,
  maxK_cos = 16,
  num_fold_final = 5
)
}
\arguments{
\item{W}{data frame with observations in the rows and baseline covariates in the columns.}

\item{A}{binary treatment vector, where 1 indicates that an individual was treated.}

\item{Y}{outcome taking values between 0 and 1. If Y is not bounded in this interval, then should rescale it before calling this function so that it is.}

\item{ngrid}{number of grid points used in approximation procedure. Selecting larger values of ngrid should yield better performance, but slower runtime.}

\item{dens_init_method}{method used to obtain initial estimates of the counterfactual density. Should be either 'ranger', in which case random forest is used, or 'kde', in which case a kernel density estimator is used.}

\item{maxK_legendre}{Maximum number of functions in the Legendre basis to consider during cross-validation.}

\item{maxK_cos}{Maximum number of functions in the cosine basis to consider during cross-validation.}

\item{num_fold_final}{Number of folds used during cross-fitting of final estimator. Should be at least 2.}
}
\value{
a list containing
\item{best_fun}{A function taking as input a vector of outcome values and outputting one-step estimates of the counterfactual density function at those values. Estimated by cross-validation to be the best of any of the estimates considered: the plug-in estimator, and one-step estimators that use the Legendre or cosine basis.}
\item{best_legendre_fun}{Same as the above, but only focusing on one-step estimators that use the Legendre basis.}
\item{best_cos_fun}{Same as the above, but only focusing on one-step estimators that use the cosine basis.}
\item{which_best}{Estimator selected for best_fun.}
\item{which_best_legendre}{Estimator selected for best_legendre_fun}
\item{which_best_cos}{Estimator selected for best_cos_fun}
}
\description{
Regularized one-step estimator of counterfactual density function under treatment A=1. Number of basis terms to include in the bias correction step is selected by cross-validation.
}
\details{
Propensity score is estimated using the ranger package.

An alternative approach for estimating this function is implemented in the npcausal R package (Kennedy et al., 2021).
}
\examples{
# sample size
n = 500

# simulate data
dat = sim_data(n,setting='zeroBothSides')
W = dat$W
A = dat$A
Y = dat$Y

cv_out = cv_density(W,A,Y,ngrid=500,num_fold_final=2)
cv_out$which_best
ygrid = seq(0.0025,0.9975,by=0.0025)
plot(ygrid,cv_out$best_fun(ygrid),xlab='y',ylab='Counterfactual Density Estimate',type='l')
}
\references{
A. Luedtke and I. Chung, ``One-Step Estimation of Differentiable Hilbert-Valued Parameters,'' \emph{arXiv}, 2023.

E. H. Kennedy, S. Balakrishnan, and L. Wasserman, ``Semiparametric counterfactual density estimation,'' \emph{arXiv}, 2021.
}
