% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{density_test}
\alias{density_test}
\title{Test for equality of counterfactual densities}
\usage{
density_test(
  W,
  A,
  Y,
  ngrid = 500,
  dens_init_method = "ranger",
  maxK_legendre = 50,
  maxK_cos = 200,
  num_fold = 2,
  num_boot = 2000,
  beta_params = c(2, 5),
  lambda = 0.5
)
}
\arguments{
\item{W}{data frame with observations in the rows and baseline covariates in the columns.}

\item{A}{binary treatment vector, where 1 indicates that an individual was treated.}

\item{Y}{outcome taking values between 0 and 1. If Y is not bounded in this interval, then should rescale it before calling this function so that it is.}

\item{ngrid}{number of grid points used in approximation procedure. Selecting larger values of ngrid should yield better performance, but slower runtime.}

\item{dens_init_method}{method used to obtain initial estimates of the two counterfactual densities. Should be either 'ranger', in which case random forest is used, or 'kde', in which case a kernel density estimator is used.}

\item{maxK_legendre}{Number of functions in the Legendre basis to use when constructing the test. In practice we've observed numerical instability when this value is taken much larger than 50.}

\item{maxK_cos}{Number of functions in the cosine basis to use when constructing the test. We have not observed any numerical instability when this value is taken large, though selecting very large values will increase the runtime and memory requirements.}

\item{num_fold}{Number of folds used during cross-fitting. Should be at least 2.}

\item{num_boot}{Number of bootstrap replications used to calculate p-value.}

\item{beta_params}{Regularization parameters used when defining the test statistic. The k-th term of the basis expansion is weighted by \code{1/(1+(k/beta_params[2])^beta_params[1])}. Test only asymptotically valid if \code{beta_params[2]>0} and \code{beta_params[1]>1/2}.}

\item{lambda}{Parameter used to define the standardization operator used to define the correlation-based Wald-type test. Should take a strictly positive value that's less than 1. Values too close to zero may lead to numerical instability, and values close to 1 will lead a test that's similar to the spherical test. See the reference for details.}
}
\value{
The p-values of the spherical and Wald-type tests based on the cosine and Legendre bases.
}
\description{
Test for the equality of counterfactual densities under treatments A=1 and A=0.
}
\details{
Defined based on a one-step estimator using the regularized spherical or Wald-type test defined in the reference. These tests either take a Wald-type or spherical form, and defining them requires selecting an orthonormal basis of L2(\code{[0,1]}). Here, we report results for the cosine basis and a transformation of the Legendre basis.

We recommend basing inference on the Wald-type test as defined with the cosine basis.

Propensity score is estimated using the ranger package.
}
\examples{
# sample size
n = 250

# simulate data from alternative
dat = sim_data(n,setting='nonzeroBothSides',cos_parameter=1)
W = dat$W
A = dat$A
Y = dat$Y

density_test(W,A,Y,ngrid=500,num_fold=2,num_boot=2000)


# simulate data from null
dat = sim_data(n,setting='nonzeroBothSides',cos_parameter=0)
W = dat$W
A = dat$A
Y = dat$Y

density_test(W,A,Y,ngrid=500,num_fold=2,num_boot=2000)
}
\references{
A. Luedtke and I. Chung, ``One-Step Estimation of Differentiable Hilbert-Valued Parameters,'' \emph{arXiv}, 2023.
}
