% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{mmd_test}
\alias{mmd_test}
\title{Maximum mean discrepancy between counterfactual distributions}
\usage{
mmd_test(W, A, Y, ngrid = 500, num_fold = 2, num_boot = 2000, bw_mult = 1)
}
\arguments{
\item{W}{data frame with observations in the rows and baseline covariates in the columns.}

\item{A}{binary treatment vector, where 1 indicates that an individual was treated.}

\item{Y}{outcome taking values between 0 and 1. If Y is not bounded in this interval, then should rescale it before calling this function so that it is.}

\item{ngrid}{number of grid points used in approximation procedure. Selecting larger values of ngrid should yield better performance, but slower runtime.}

\item{num_fold}{Number of folds used during cross-fitting. Should be at least 2.}

\item{num_boot}{Number of bootstrap replications used to calculate p-value.}

\item{bw_mult}{The reported MMD is based on the Gaussian kernel with bandwidth equal to the median heuristic -- namely, the median of the off-diagonal elements of the distance matrix dist(Y) -- times bw_mult.}
}
\value{
The p-value of the test.
}
\description{
Maximum mean discrepancy between counterfactual distributions under treatments A=1 and A=0.
}
\details{
Maximum mean discrepancy estimator defined by computing the distance between the one-step estimators of the counterfactual kernel mean embeddings under A=1 and A=0.

All needed nuisances are estimated using the ranger package.
}
\examples{
# sample size
n = 250

# simulate data from alternative
dat = sim_data(n,setting='nonzeroBothSides',cos_parameter=1)
W = dat$W
A = dat$A
Y = dat$Y

mmd_test(W,A,Y,ngrid=500,num_fold=2,num_boot=2000,bw_mult=1)


# simulate data from null
dat = sim_data(n,setting='nonzeroBothSides',cos_parameter=0)
W = dat$W
A = dat$A
Y = dat$Y

mmd_test(W,A,Y,ngrid=500,num_fold=2,num_boot=2000,bw_mult=1)
}
\references{
A. Luedtke and I. Chung, ``One-Step Estimation of Differentiable Hilbert-Valued Parameters,'' \emph{arXiv}, 2023.

J. Fawkes, R. Hu, R. J. Evans, and D. Sejdinovic, ``Doubly Robust Kernel Statistics for Testing Distributional Treatment Effects Even Under One Sided Overlap,'' \emph{arXiv}, 2022.

M. Muandet, M. Kanagawa, S.  Saengkyongam, and S. Marukatat, ``Counterfactual mean embeddings,'' \emph{J Mach Learn Res}, 2021.
}
