% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data from distributions without bandlimited counterfactual density}
\usage{
sim_data(n, setting = "nonzeroBothSides", cos_parameter = 1)
}
\arguments{
\item{n}{sample size}

\item{setting}{one of \code{'spikeLeft'}, \code{'nonzeroBothSides'}, and \code{'zeroBothSides'}.}

\item{cos_parameter}{Nonnegative integer used to define distribution of the counterfactual density under A=0.}
}
\value{
a list containing
\item{W}{data frame with observations in the rows and baseline covariates in the columns.}
\item{A}{binary treatment vector, where 1 indicates that an individual was treated.}
\item{Y}{vector of outcomes.}
\item{Y0}{vector of counterfactual outcomes under A=0.}
\item{Y1}{vector of counterfactual outcomes under A=1.}
}
\description{
Simulates data from counterfactual density scenarios considered in Luedtke and Chung (2023).
}
\details{
The counterfactual density under A=1 is defined via the \code{setting} input. Details on the three scenarios listed here can be found in Luedtke and Chung (2023).

The counterfactual density under A=0 is defined via the \code{cos_parameter} input. When \code{cos_parameter=0}, the counterfactual density under A=0 is the same as that under A=1. When \code{cos_parameter} is equal to \code{k>0} and \code{setting='nonzeroBothSides'}, data are simulated from the scenario labeled 'Alt k' in Luedtke and Chung (2023).
}
\examples{
sim_data(100)
}
\references{
A. Luedtke and I. Chung, ``One-Step Estimation of Differentiable Hilbert-Valued Parameters,'' \emph{arXiv}, 2023.
}
