% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{sim_data_bandlimited}
\alias{sim_data_bandlimited}
\title{Simulate data from distribution with bandlimited counterfactual density}
\usage{
sim_data_bandlimited(n)
}
\arguments{
\item{n}{sample size}
}
\value{
a list containing
\item{W}{data frame with observations in the rows and baseline covariates in the columns.}
\item{A}{binary treatment vector, where 1 indicates that an individual was treated.}
\item{Y}{vector of outcomes.}
\item{Y0}{vector of counterfactual outcomes under A=0.}
\item{Y1}{vector of counterfactual outcomes under A=1.}
}
\description{
Simulates data from bandlimited counterfactual density scenario considered in Luedtke and Chung (2023).
}
\details{
The Fourier transform of the counterfactual density under A=1 has support \code{[-2,2]}.
}
\examples{
sim_data_bandlimited(100)
}
\references{
A. Luedtke and I. Chung, ``One-Step Estimation of Differentiable Hilbert-Valued Parameters,'' \emph{arXiv}, 2023.
}
